+++ internal/machine/config.go
@@ -75,3 +75,18 @@ func ConfigCompare(ctx context.Context, original api.MachineConfig, new api.Mach
 
 	return ""
 }
+
+func CloneConfig(orig api.MachineConfig) (*api.MachineConfig, error) {
+	config := &api.MachineConfig{}
+
+	data, err := json.Marshal(orig)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := json.Unmarshal(data, config); err != nil {
+		return nil, err
+	}
+
+	return config, err
+}
