+++ internal/machine/update.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/flaps"
+	"github.com/superfly/flyctl/internal/prompt"
 	"github.com/superfly/flyctl/internal/watch"
 	"github.com/superfly/flyctl/iostreams"
 )
@@ -41,4 +42,33 @@ func Update(ctx context.Context, m *api.Machine, input *api.LaunchMachineInput, 
 	fmt.Fprintf(io.Out, "Machine %s updated successfully!\n", colorize.Bold(m.ID))
 
 	return nil
+}
+
+func ConfirmUpdate(ctx context.Context, machine *api.Machine, targetConfig api.MachineConfig) (bool, error) {
+	var (
+		io       = iostreams.FromContext(ctx)
+		colorize = io.ColorScheme()
+	)
+
+	diff := ConfigCompare(ctx, *machine.Config, targetConfig)
+	if diff == "" {
+		return true, nil
+	}
+	fmt.Fprintf(io.Out, "Configuration changes to be applied to machine: %s.\n", colorize.Bold(machine.ID))
+	fmt.Fprintf(io.Out, "%s\n", diff)
+
+	const msg = "Apply changes?"
+
+	switch confirmed, err := prompt.Confirmf(ctx, msg); {
+	case err == nil:
+		if !confirmed {
+			return false, nil
+		}
+	case prompt.IsNonInteractive(err):
+		return false, prompt.NonInteractiveError("auto-confirm flag must be specified when not running interactively")
+	default:
+		return false, err
+	}
+
+	return true, nil
 }
